Hi guys :) I finally got something to show. :) This is just a demo of the cover system, and doesn't have the vertex cover thing in it or bullets shooting.
The reason it doesn't is, because of ai path finding, and stuff. We need to get all that stuff in the system before I add in vertex covering system that works with this cover system. 

The reason it took me so long was trying to get theory on how I would program a cover system that would not only work  for a fast fps, but will also work with a fps stealth game. :)

I also add a collision system in there to see how the ai behave for the demo. :)